Refined GPS tracking, session management, and DAO data types
This commit improves the accuracy of drag racing and track session calculations, adds permission safety to GPS providers, and standardizes ID types in DAOs.

**Key Changes:**

*   **`DragTimeCalculation.kt`:**
    *   Improved quarter-mile accuracy by measuring distance relative to the run's start (`runStartDistanceMeters`) rather than the total session distance.
    *   Ensures `isReadyForRun` is reset when tracking is reset.
*   **`TrackGeometry.kt`:**
    *   Refined line-crossing logic to only register "forward" crossings (`ENTERING` direction). This prevents duplicate lap triggers if a vehicle rolls backward over a start/finish line.
*   **`PhoneGpsProvider.kt`:**
    *   Added explicit permission checks for `ACCESS_FINE_LOCATION` and `ACCESS_COARSE_LOCATION` before starting location updates to prevent crashes or silent failures.
*   **`TimeAttackViewModel.kt`:**
    *   Simplified session ID assignment by directly using the return value of `sessionManager.startSession()`.
*   **DAOs (`RawGPSDataDao.kt`, `TrackCoordinatesDataDAO.kt`):**
    *   Updated `delete` methods to use `Long` for `sessionId` and `trackId` to maintain consistency with the rest of the database schema.

+++ app/src/main/java/com/example/trackpro/dao/RawGPSDataDao.kt
@@ -18,5 +18,5 @@ interface RawGPSDataDao {
     suspend fun getGPSDataBySession(sessionId: Long?): List<RawGPSData>
 
     @Query("DELETE FROM raw_gps_data WHERE sessionId = :sessionId")
-    suspend fun deleteGPSDataBySession(sessionId: Int)
+    suspend fun deleteGPSDataBySession(sessionId: Long)
 }
